+++ RELEASE_NOTES.md
@@ -24,9 +24,15 @@
          *    1 DDR controller implemented in the SH (always available)
          *    3 DDR controllers implemented in the CL (configurable number of implemented controllers allowed)
 
+## Release 1.4.13 (See [ERRATA](./ERRATA.md) for unsupported features)
+* FPGA developer kit now supports Xilinx Vivado/Vitis 2019.2
+* To upgrade, use [Developer AMI v1.8.0](https://aws.amazon.com/marketplace/pp/B06VVYBLZZ) on the AWS Marketplace.
+       
 ## Release 1.4.12 (See [ERRATA](./ERRATA.md) for unsupported features)
-* Added support for Xilinx tool versions released in the AWS China Marketplace.
-* Updated NICE DCV setup instructions.
+* Added supported versions for BJS AMI's
+* Added link to the re:Invent 19 F1 workshop
+* Fixed missing extern C declaration by PR #473 
+* Documentation Path fixes from PR #466, #468 and #470 
 
 ## Release 1.4.11 (See [ERRATA](./ERRATA.md) for unsupported features)
 * FPGA developer kit now supports Xilinx SDx/Vivado 2019.1